requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-rest-framework==0.1.0
django-smartfields==1.1.3
djangorestframework==3.12.2
Pillow==8.1.2
pkg-resources==0.0.0
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
